# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_photo_api_v1_vehicles__vehicle_id__photos_post": {
        "properties": {
          "angle": {
            "title": "Angle",
            "type": "string"
          },
          "captured_at": {
            "title": "Captured At",
            "type": "string"
          },
          "checksum_sha256": {
            "title": "Checksum Sha256",
            "type": "string"
          },
          "client_uuid": {
            "format": "uuid",
            "title": "Client Uuid",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "inspection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Id"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "client_uuid",
          "angle",
          "phase",
          "captured_at",
          "checksum_sha256",
          "width",
          "height"
        ],
        "title": "Body_upload_photo_api_v1_vehicles__vehicle_id__photos_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PhotoRead": {
        "description": "`url` est calculée à la volée par le backend de stockage actif\n(`app/services/storage/service.py`) — jamais stockée en base.",
        "properties": {
          "angle": {
            "title": "Angle",
            "type": "string"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "client_uuid": {
            "format": "uuid",
            "title": "Client Uuid",
            "type": "string"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inspection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Id"
          },
          "is_placeholder": {
            "title": "Is Placeholder",
            "type": "boolean"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "upload_state": {
            "title": "Upload State",
            "type": "string"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "work_order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Order Id"
          }
        },
        "required": [
          "id",
          "vehicle_id",
          "inspection_id",
          "work_order_id",
          "angle",
          "phase",
          "content_type",
          "byte_size",
          "width",
          "height",
          "client_uuid",
          "upload_state",
          "is_placeholder",
          "captured_at",
          "uploaded_at",
          "url"
        ],
        "title": "PhotoRead",
        "type": "object"
      },
...
      "RequiredAnglesResponse": {
        "description": "Parcours d'angles imposé (brief J2) — ce que le front doit encore capturer pour un\n`inspection_id` donné.",
        "properties": {
          "captured_angles": {
            "items": {
              "type": "string"
            },
            "title": "Captured Angles",
            "type": "array"
          },
          "missing_angles": {
            "items": {
              "type": "string"
            },
            "title": "Missing Angles",
            "type": "array"
          },
          "required_angles": {
            "items": {
              "type": "string"
            },
            "title": "Required Angles",
            "type": "array"
          }
        },
        "required": [
          "required_angles",
          "captured_angles",
          "missing_angles"
        ],
        "title": "RequiredAnglesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Outil interne de gestion d'achat de véhicules d'occasion — démonstration.",
    "title": "Cardan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/photos/file/{bucket}/{key}": {
      "get": {
        "operationId": "get_photo_file_api_v1_photos_file__bucket___key__get",
        "parameters": [
          {
            "in": "path",
            "name": "bucket",
            "required": true,
            "schema": {
              "title": "Bucket",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Photo File",
        "tags": [
          "photos"
        ]
      }
    },
...
    "/api/v1/vehicles/{vehicle_id}/photos": {
      "get": {
        "operationId": "list_vehicle_photos_api_v1_vehicles__vehicle_id__photos_get",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vehicle Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inspection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inspection Id"
            }
          },
          {
            "in": "query",
            "name": "phase",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phase"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PhotoRead"
                  },
                  "title": "Response List Vehicle Photos Api V1 Vehicles  Vehicle Id  Photos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Vehicle Photos",
        "tags": [
          "photos"
        ]
      },
      "post": {
        "operationId": "upload_photo_api_v1_vehicles__vehicle_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vehicle Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_v1_vehicles__vehicle_id__photos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photo",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/vehicles/{vehicle_id}/photos/required-angles": {
      "get": {
        "description": "Ce que le front doit encore capturer pour `inspection_id` — la liste des 12 angles\nimposés est portée par le backend (`REQUIRED_PHOTO_ANGLES`), jamais recopiée côté front.",
        "operationId": "get_required_angles_api_v1_vehicles__vehicle_id__photos_required_angles_get",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vehicle Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequiredAnglesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Required Angles",
        "tags": [
          "photos"
        ]
      }
    },
...
  }
}